Rename load action and constants modules to 'entry'
